add damage fix, menu changes

+++ core/src/main/java/com/shooter/weapons/Weapon.java
@@ -9,14 +9,14 @@ import java.util.ArrayList;
 
 public abstract class Weapon {
     public String name;
-    public int damage;
+    public float damage;
     public int capacity;
     public int currentCapacity;
     public Texture bulletTexture;
 
-    public Weapon(String name, int damage, int cap) {
+    public Weapon(String name, float damage, int cap) {
         this.name = name;
-        this.damage = damage;
+        this.damage = damage ;
         this.capacity = cap;
         this.currentCapacity = cap;
         bulletTexture = new Texture("player/bullet.png");
